+++ src/main/java/com/ward/ward_server/api/user/controller/AuthController.java
@@ -15,7 +15,6 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/auth")
 public class AuthController {
     private final AuthService authService;
-//    private final UserService userService;
     private final JwtProperties properties;
 
     //TODO 현재 Email 만 같으면 로그인 처리 되게 되어있음.
@@ -31,27 +30,26 @@ public class AuthController {
         return ApiResponse.ok(tokens);
     }
 
-    //TODO 닉네임 중복 에러 반환
     @PostMapping
     public ApiResponse<JwtTokens> register(@RequestBody @Validated RegisterRequest request) {
         JwtTokens tokens = authService.registerUser(request);
         return ApiResponse.ok(tokens);
     }
 
-    //TODO 관리자 권한 부여 방식 토의 필요
-    //관리자 권한 부여
-//    @PutMapping("/grantAdmin")
-//    public ApiResponse<Void> grantAdminRole(@RequestBody @Validated RoleChangeRequest request) {
-//        userService.grantAdminRole(request.userId());
-//        return ApiResponse.ok();
-//    }
-//
-//    // 사용자 권한 부여
-//    @PutMapping("/grantUser")
-//    public ApiResponse<Void> grantUserRole(@RequestBody @Validated RoleChangeRequest request) {
-//        userService.grantUserRole(request.userId());
-//        return ApiResponse.ok();
-//    }
+    // TODO 관리자 권한 부여 방식 토의 필요
+    // 관리자 권한 부여
+    // @PutMapping("/grantAdmin")
+    // public ApiResponse<Void> grantAdminRole(@RequestBody @Validated RoleChangeRequest request) {
+    //     userService.grantAdminRole(request.userId());
+    //     return ApiResponse.ok();
+    // }
+
+    // 사용자 권한 부여
+    // @PutMapping("/grantUser")
+    // public ApiResponse<Void> grantUserRole(@RequestBody @Validated RoleChangeRequest request) {
+    //     userService.grantUserRole(request.userId());
+    //     return ApiResponse.ok();
+    // }
 
     @GetMapping("/checkNickname")
     public ApiResponse<Boolean> checkDuplicateNickname(@RequestParam("nickname") String nickname) {
